refactor(core): optimize setup by moving dependency check to setup.ts
